refactor: refactored core logic, and improved performance.
