+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "grafana-dynamic-search",
-  "version": "1.4.1",
+  "version": "1.5.0",
   "scripts": {
     "build": "webpack -c ./.config/webpack/webpack.config.ts --env production",
     "dev": "webpack -w -c ./.config/webpack/webpack.config.ts --env development",
