+++ src/utils.spec.ts
@@ -93,54 +93,46 @@ describe('utils', () => {
 
     it('should return original values if no regex provided', () => {
       const result = applyRegexTransform(values, null);
-      expect(result).toEqual([
-        { label: 'node-01', value: 'node-01', description: undefined },
-        { label: 'node-02', value: 'node-02', description: undefined },
-        { label: 'other', value: 'other', description: undefined },
-      ]);
+      expect(result).toEqual(values);
     });
 
     it('should apply regex capture group', () => {
       const regex = /node-(\d+)/;
       const result = applyRegexTransform(values, regex);
       expect(result).toEqual([
-        { label: '01', value: '01', description: undefined },
-        { label: '02', value: '02', description: undefined },
-        { label: 'other', value: 'other', description: undefined },
+        { text: '01', value: '01', __originalText: 'node-01' },
+        { text: '02', value: '02', __originalText: 'node-02' },
+        { text: 'other' },
       ]);
     });
 
     it('should return original value if regex does not match', () => {
       const regex = /nomatch/;
       const result = applyRegexTransform(values, regex);
-      expect(result).toEqual([
-        { label: 'node-01', value: 'node-01', description: undefined },
-        { label: 'node-02', value: 'node-02', description: undefined },
-        { label: 'other', value: 'other', description: undefined },
-      ]);
+      expect(result).toEqual(values);
     });
 
     it('should handle complex regex', () => {
        const regex = /(.*)/;
        const result = applyRegexTransform(values, regex);
        expect(result).toEqual([
-         { label: 'node-01', value: 'node-01', description: undefined },
-         { label: 'node-02', value: 'node-02', description: undefined },
-         { label: 'other', value: 'other', description: undefined },
+         { text: 'node-01', value: 'node-01', __originalText: 'node-01' },
+         { text: 'node-02', value: 'node-02', __originalText: 'node-02' },
+         { text: 'other', value: 'other', __originalText: 'other' },
        ]);
     });
 
     it('should handle undefined text gracefully', () => {
         const valuesWithUndefined = [{ text: undefined }] as unknown as MetricFindValue[];
         const result = applyRegexTransform(valuesWithUndefined, null);
-        expect(result).toEqual([{ label: '', value: '', description: undefined }]);
+        expect(result).toEqual(valuesWithUndefined);
     });
 
     it('should handle undefined text with regex gracefully', () => {
         const valuesWithUndefined = [{ text: undefined }] as unknown as MetricFindValue[];
         const regex = /test/;
         const result = applyRegexTransform(valuesWithUndefined, regex);
-        expect(result).toEqual([{ label: '', value: '', description: undefined }]);
+        expect(result).toEqual(valuesWithUndefined);
     });
 
     it('should preserve description field', () => {
@@ -149,8 +141,7 @@ describe('utils', () => {
             { text: 'node-02' },
         ] as unknown as MetricFindValue[];
         const result = applyRegexTransform(valuesWithDescription, null);
-        expect(result[0].description).toBe('First node');
-        expect(result[1].description).toBeUndefined();
+        expect(result).toEqual(valuesWithDescription);
     });
 
     it('should preserve description field with regex transform', () => {
@@ -159,7 +150,7 @@ describe('utils', () => {
         ] as unknown as MetricFindValue[];
         const regex = /node-(\d+)/;
         const result = applyRegexTransform(valuesWithDescription, regex);
-        expect(result[0]).toEqual({ label: '01', value: '01', description: 'First node' });
+        expect(result[0]).toEqual({ text: '01', value: '01', description: 'First node', __originalText: 'node-01' });
     });
   });
